Update Documentation for Table
