+++ docs/src/app/components/pages/components/Table/README.md
@@ -1,0 +1,5 @@
+## Table
+[Tables](https://www.google.com/design/spec/components/data-tables.html#)
+are used to display data and to organize it.
+
+### Examples
