+++ src/table/table-footer.jsx
@@ -7,7 +7,18 @@ import ThemeManager from '../styles/theme-manager';
 const TableFooter = React.createClass({
 
   propTypes: {
+    /**
+     * Controls whether or not header rows should be adjusted
+     * for a checkbox column. If the select all checkbox is true,
+     * this property will not influence the number of columns.
+     * This is mainly useful for "super header" rows so that
+     * the checkbox column does not create an offset that needs
+     * to be accounted for manually.
+     */
     adjustForCheckbox: React.PropTypes.bool,
+    /**
+     * Children passed to table footer.
+     */
     children: React.PropTypes.node,
 
     /**
